added paginate file, standarized code
