+++ controllers/index.js
@@ -1,10 +1,11 @@
-const createHttpError = require('http-errors')
-const { Test } = require('../database/models')
-const { endpointResponse } = require('../helpers/success')
-const { catchAsync } = require('../helpers/catchAsync')
+const createHttpError = require("http-errors");
+const { Test } = require("../database/models");
+const { endpointResponse } = require("../helpers/success");
+const { catchAsync } = require("../helpers/catchAsync");
 
 // example of a controller. First call the service, then build the controller method
 module.exports = {
+
   get: catchAsync(async (req, res, next) => {
     try {
       const response = await Test.findAll()
@@ -21,4 +22,4 @@ module.exports = {
       next(httpError)
     }
   }),
-}
+};
